chore: move components into folder, small ui imprv
