+++ src/components/index.js
@@ -1,0 +1,4 @@
+export { AddTask } from './AddTask'
+export { Task } from './Task'
+export { TaskList } from './TaskList'
+export { MemoedTotalCount } from './Total'
